Wait for sd-workstation VMs to shutdown
This is a follow-up to 1533363cf7ce7e, which doesn't seem to have solved
the issue we were facing. My suspicion is that the test runner tried to
access the VMs before they shut down, so they didn't.

Just in case the VMs don't shut down in 60 seconds, we abort so it isn't
indefinitely waiting.

Fixes #47.