Provide context to Slack as to why the build ran.
This starts by passing --context to the begin.py (in sd-dev), which
then passes it to dom0 as part of the RPC qrexec argument.

dom0 then passes the context back to sd-dev in the call to status.py,
which allows sd-dev to populate the context in the Slack notification
that it sends.

We can therefore pass --context in nightlies.py to the run.py on
the bastion which passes this on to sd-dev (by way of a qvm-run in
dom0) once the Qubes machine is booted.

Similarly we will be able to pass --context when triggering run.py
via the Github webhook on git push events.