Don't use cached qubesadmin.Qubes() object
qubesadmin caches the list of VMs, so instantiating the object before
provisioning means using it afterwards will not return the newly created
VMs.

Digging a bit deeper, the original use of qubesadmin (to teardown the
sys-usb VM) was removed in 86fe060 so that load can be removed entirely.

Then just instantiate a `Qubes()` object in our shutdown function and
don't persist it for any longer.

Refs #47.