Notify GitHub build is queued from the bastion
Currently the "begin" status is only reported to GitHub once the Qubes
VM starts. So for the period of git push --> VM start, nothing is
displayed on GitHub that we're in a pending state, which can easily be
tens of minutes if it's waiting for a VM.

We now notify GitHub at the beginning of run.py to add the status with a
message of "The build is queued." instead of in begin.py.

Fixes #63.